Core/Cleanup: * move repeating code into separate method for storing locale strings in ObjectMgr (copy/paste sucks, you know) * fix 'signed/unsigned' warnings (at least in VS) * fix some other warnings and cleanup relative code
--HG--
branch : trunk

+++ src/server/game/Entities/Unit/StatSystem.cpp
@@ -40,7 +40,7 @@ inline bool _ModifyUInt32(bool apply, uint32& baseValue, int32& amount)
     else 
     {
         // Make sure we do not get uint32 overflow.
-        if (amount > baseValue)
+        if (amount > int32(baseValue))
             amount = baseValue;
         baseValue -= amount;
     }
